Time: 1 ms (99.97%), Space: 42.6 MB (63.48%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.java
@@ -1,6 +1,6 @@
 class Solution {
     public int reverse(int x) {
-        double rev = 0;
+        long rev = 0;
 
         while(x!=0){
             int digit = x % 10;
